create common body
